# [DNM] apd: embed small coefficient values in Decimal struct