@nvanbenschoten 2022-01-04
## Updates

### Integration into CRDB

I took a stab at integrating this into CRDB in [this nightmare of a prototype](https://github.com/nvanbenschoten/cockroach/commit/83e9dad8712b6453cccb4161804b8bf8769cc9f4).

The restriction that `apd.Decimal` can no longer be copied by value wasn't too much of a hurdle in most places. In fact, it actually forced some improvements in the key and value decoding code, as it prevented memcpys and wasted allocations by preventing us from writing code that caused them. For example, `DecodeDecimalAscending`'s signature changed from:
```
func DecodeDecimalAscending(buf []byte, tmp []byte) ([]byte, apd.Decimal, error)
```
to
```
func DecodeDecimalAscending(dec *apd.Decimal, buf []byte, tmp []byte) ([]byte, error)
```
This in turn forces the code to be slightly more deliberate about where it is decoding `Decimal`s into, so we end up using the `rowenc.DatumAlloc` more reliably. All in all, I think that's fine.

The more complex part of the change was in the vectorized execution engine. The complexity wasn't even entirely due to this change, but instead due to the need to first address [this todo](https://github.com/cockroachdb/cockroach/blob/798a3e544d7ed38c5ea21085054b97cac067f56d/pkg/col/coldata/native_types.go#L109-L110):
```
// TODO(yuzefovich): consider whether Get and Set on Decimals should operate on
// pointers to apd.Decimal.
```
Since we can no longer copy `Decimal`s by value, the handling of `Decimal` in the vectorized execution needs to become even more custom. This is because we don't currently have any data types that the vectorized execution engine interacts with by reference.

I spent a while learning about and playing with `execgen`. I made some progress, but eventually gave up and made changes manually. I'm sure someone who actually knows how `execgen` works would have more luck, but I really struggled with the conditional templating.

After making that change and getting CRDB compiling, I took it for a spin. First, I confirmed that we see the same reduction in heap allocations that we saw in https://github.com/cockroachdb/cockroach/pull/74369. Next, I played around a bit with the tpcds dataset, which includes many DECIMAL columns. I didn't run any tpcds queries, but I did do a few custom full-table aggregations over the DECIMAL columns in ~2GB large tables.

Nothing was scientific, but there did seem to be something around a 5% speedup compared to master on these aggregations. Or maybe I just wanted there to be. Again, not scientific.

----

### More compact memory representation

After writing this change I looked into how other systems like Postgres, MySQL, SQLServer, and Materialize handle Decimals. It turns out that only Postgres (not even Materialize, which strives for close to full PG support) supports arbitrary precision decimals (100k+ precision). Most other systems cap the precision of decimals at somewhere around 40, meaning that they don't even need to support a variable-length memory representation. They can instead inline the entire value in a single 24 byte large struct.

We do want to maintain compatibility with PG and so I think we need to continue to support arbitrary precision, but this made me question why we were tailoring so much of this code to arbitrary precision. Specifically, I started questioning why we were spending so many bytes of the `Decimal` memory representation to support absurdly large precision value. Out of the 64 bytes in a `Decimal`, 32 were going to the `big.Int` (bool + slice header) and 8 were going to the self-pointer to dynamically detect value copies which are unsafe as a second-order consequence of supporting arbitrary precision.

I drafted another prototype that further optimizes for small values at the expense of large values:  https://github.com/nvanbenschoten/apd/commit/654be62f74b7189b16c929bdb561f86a878d71f1.

The key idea is that instead of a memory representation that looks like:
```go
type Decimal struct {
    Form     int8
    Negative bool
    Exponent int32
    Coeff    BigInt {
        _inner big.Int {
            neg bool
            abs []uint
        }
        _inline [2]uint
        _addr   uintptr
    }
} // sizeof = 64 bytes
```

We have a memory representation that looks like:
```go
type Decimal struct {
    Form     int8
    Negative bool
    Exponent int32
    Coeff    BigInt {
        _inner *big.Int
        _inline [2]uint
    }
} // sizeof = 32
```

There are a few tradeoffs here:

Pros:
- reduces memory size from 64 bytes to 32 bytes, allowing us to pack two `Decimals` into a single cache line
- removes the self-referential pointer, making the Decimal struct copyable again

Cons:
- two heap allocations for large values instead of one (allocate `big.Int` then allocate large backing array)
- need to inflate stack-allocated `big.Int` on demand during arithmetic if we wanted to continue to defer all arithmetic to `big.Int`

This second con leads to code that looks like:
```go
func (z *BigInt) And(x, y *BigInt) *BigInt {
	var tmp1, tmp2, tmp3 big.Int
	zi := z.inner(&tmp1)
	zi.And(x.inner(&tmp2), y.inner(&tmp3))
	z.updateInner(zi)
	return z
}
```

This approach compared favorably to the `master` implementation on benchmarks. For instance, the reduction in heap allocations (see PR description) is mostly the same between the two prototype approaches. For the most part, this still led to net speedups across the board compared to `master`.

However, even with the reduction in memory size and on datasets large enough to not fit entirely in my CPU's cache, benchmarks did not present this new approach favorably compared to the approach taken by this PR. It appears that in this performance sensitive code, even with everything inlined, the on-demand inflation and deflation of stack-allocated `big.Int` structs on every operation was too expensive. Here's the comparison between this PR (old) and the approach taken by the [compact representation prototype](https://github.com/nvanbenschoten/apd/commit/654be62f74b7189b16c929bdb561f86a878d71f1#diff-17b95c705c56bfb96739a18a4e13e3d65a8b185f633dbafc27543e55db390cebR213) (new):
```
name                 old time/op    new time/op    delta
GDA/base-10             113µs ± 1%     114µs ± 1%    +0.51%  (p=0.005 n=10+10)
GDA/plus-10            38.3µs ± 0%    41.9µs ± 0%    +9.30%  (p=0.000 n=9+10)
GDA/divideint-10       23.1µs ± 0%    25.7µs ± 1%   +11.03%  (p=0.000 n=10+10)
GDA/remainder-10       41.7µs ± 0%    48.2µs ± 0%   +15.64%  (p=0.000 n=9+9)
GDA/minus-10           8.95µs ± 0%   10.82µs ± 0%   +20.88%  (p=0.000 n=10+10)
GDA/subtract-10         112µs ± 0%     136µs ± 0%   +21.82%  (p=0.000 n=10+9)
GDA/exp-10              130ms ± 0%     158ms ± 0%   +22.03%  (p=0.000 n=10+10)
GDA/add-10              655µs ± 0%     800µs ± 0%   +22.12%  (p=0.000 n=10+10)
GDA/comparetotal-10    24.4µs ± 1%    30.1µs ± 1%   +23.33%  (p=0.000 n=10+10)
GDA/abs-10             7.11µs ± 0%    8.81µs ± 0%   +23.79%  (p=0.000 n=10+10)
GDA/tointegralx-10     22.6µs ± 0%    28.1µs ± 0%   +24.59%  (p=0.000 n=10+7)
GDA/tointegral-10      22.2µs ± 0%    27.7µs ± 0%   +24.75%  (p=0.000 n=10+9)
GDA/multiply-10        56.3µs ± 0%    71.8µs ± 0%   +27.62%  (p=0.000 n=10+8)
GDA/divide-10           377µs ± 0%     486µs ± 0%   +28.87%  (p=0.000 n=10+10)
GDA/randoms-10         2.29ms ± 0%    2.96ms ± 0%   +29.00%  (p=0.000 n=10+9)
GDA/reduce-10          14.2µs ± 0%    18.4µs ± 0%   +29.25%  (p=0.000 n=10+10)
GDA/log10-10            114ms ± 0%     148ms ± 0%   +29.61%  (p=0.000 n=10+9)
GDA/quantize-10        85.0µs ± 0%   110.1µs ± 0%   +29.62%  (p=0.000 n=10+10)
GDA/ln-10              90.1ms ± 0%   117.0ms ± 0%   +29.82%  (p=0.000 n=10+9)
GDA/compare-10         34.5µs ± 1%    44.9µs ± 1%   +30.18%  (p=0.000 n=10+10)
GDA/power-10            233ms ± 0%     304ms ± 0%   +30.39%  (p=0.000 n=9+10)
GDA/squareroot-10      28.7ms ± 0%    37.4ms ± 0%   +30.45%  (p=0.000 n=8+9)
GDA/rounding-10         492µs ± 0%     645µs ± 0%   +31.13%  (p=0.000 n=10+9)
GDA/cuberoot-apd-10    2.14ms ± 0%    2.81ms ± 0%   +31.14%  (p=0.000 n=10+9)
GDA/powersqrt-10        433ms ± 0%     571ms ± 0%   +31.96%  (p=0.000 n=10+10)
```

So I think we'll want to stick with this PR.