@nvanbenschoten 2022-01-05
Thanks for taking a look, Yahor. I agree with everything you said. The no-copy limitation appeared to make the integration into the vectorized execution engine difficult for a number of different reasons, mostly stemming from the need to operate on references instead of values. It looked possible, but would likely be a lot of work.

In light of your exploration, I took another look at the "More compact memory representation" alternative. This alternative approach is appealing because it does not use a self-referential pointer, so it can be copied by value safely (though `Decimal.Set` should still be used to avoid aliasing when that is a problem). I mentioned above that the repeat inflation and deflation of stack-allocated `big.Int` structs on every operation was too expensive. We could see this in the benchmark results when compared to this PR.

I explored what it would take to bypass `big.Int` for basic arithmetic on small values, similar to what the `ericlagergren/decimal` library does. As it turns out, the separation between `apd.Decimal` and `apd.BigInt` made this all fairly straightforward — `apd.Decimal` remains entirely oblivious of these optimizations and `apd.BigInt` can continue to fallback to `big.Int` whenever the arithmetic gets hard. Furthermore, it also means that even the complex arithmetic at the `Decimal` level (e.g. `Decimal.Sqrt`) still benefits from fast-paths on simple arithmetic, because these complex `Decimal` operations are composed of many incremental operations on `BigInt`s.

Here's the change to add these fast-paths to the compact memory representation approach: https://github.com/cockroachdb/apd/commit/c5c0544dc082aed7d47c108cc70508a96f46731d. It's not a lot of code, and it benefits from the existing `BigInt` and `Decimal` test coverage.

With that commit, the comparison between the two approaches shift in favor of the compact (and copyable) memory representation (old is this PR, new is https://github.com/cockroachdb/apd/commit/c5c0544dc082aed7d47c108cc70508a96f46731d):
```
name                 old time/op    new time/op    delta
GDA/squareroot-10      28.7ms ± 0%    15.7ms ± 0%   -45.30%  (p=0.000 n=10+10)
GDA/powersqrt-10        433ms ± 0%     240ms ± 0%   -44.45%  (p=0.000 n=10+10)
GDA/rounding-10         493µs ± 0%     297µs ± 0%   -39.68%  (p=0.000 n=10+10)
GDA/divideint-10       23.1µs ± 0%    15.3µs ± 0%   -33.58%  (p=0.000 n=10+10)
GDA/divide-10           377µs ± 0%     260µs ± 0%   -30.98%  (p=0.000 n=8+10)
GDA/remainder-10       41.7µs ± 0%    31.7µs ± 0%   -24.01%  (p=0.000 n=10+10)
GDA/tointegral-10      22.2µs ± 0%    17.7µs ± 0%   -20.29%  (p=0.000 n=10+9)
GDA/randoms-10         2.30ms ± 0%    1.85ms ± 0%   -19.51%  (p=0.000 n=10+10)
GDA/tointegralx-10     22.5µs ± 0%    18.2µs ± 0%   -19.35%  (p=0.000 n=10+10)
GDA/compare-10         34.5µs ± 1%    28.2µs ± 0%   -18.29%  (p=0.000 n=10+10)
GDA/minus-10           8.94µs ± 0%    7.47µs ± 0%   -16.39%  (p=0.000 n=9+9)
GDA/abs-10             7.13µs ± 0%    5.99µs ± 0%   -15.95%  (p=0.000 n=10+10)
GDA/subtract-10         112µs ± 0%      97µs ± 0%   -13.10%  (p=0.000 n=10+10)
GDA/cuberoot-apd-10    2.14ms ± 0%    1.89ms ± 0%   -11.91%  (p=0.000 n=8+10)
GDA/exp-10              130ms ± 0%     118ms ± 1%    -9.34%  (p=0.000 n=8+10)
GDA/ln-10              90.3ms ± 0%    82.3ms ± 0%    -8.79%  (p=0.000 n=10+10)
GDA/log10-10            114ms ± 0%     106ms ± 0%    -7.25%  (p=0.000 n=10+8)
GDA/quantize-10        85.1µs ± 0%    80.1µs ± 0%    -5.89%  (p=0.000 n=10+9)
GDA/power-10            233ms ± 0%     220ms ± 0%    -5.66%  (p=0.000 n=10+9)
GDA/reduce-10          14.3µs ± 0%    13.5µs ± 0%    -5.42%  (p=0.000 n=9+10)
GDA/plus-10            38.4µs ± 0%    37.3µs ± 0%    -2.93%  (p=0.000 n=9+10)
GDA/multiply-10        56.5µs ± 1%    55.8µs ± 1%    -1.19%  (p=0.000 n=10+10)
GDA/comparetotal-10    24.4µs ± 2%    24.4µs ± 2%      ~     (p=0.971 n=10+10)
GDA/base-10             114µs ± 1%     114µs ± 1%    +0.55%  (p=0.019 n=10+10)
GDA/add-10              657µs ± 0%     674µs ± 0%    +2.61%  (p=0.000 n=10+10)
```

_(`GDA/add` is a somewhat unrealistic benchmark because almost all arithmetic takes place on high-precision values that aren't inlined by either approach. `GDA/minus` contains a more realistic distribution)_

I'll look into packaging up that alternative approach and addressing the existing comments of yours that apply to it.